[TRTLLM-12341][fix] Fix default-config enc-dec crashes and Whisper front-end robustness
- Cap enc-dec warmup dummy length by max_target_positions (OOB decoder
  position-table gather silently corrupted transcripts)
- Disable KV-cache estimation for enc-dec (text-only estimation dummies
  break the cross-KV block accounting)
- Disable CUDA graphs for fp32 enc-dec (undersized cross-attention
  workspace crashes cuBLAS during capture)
- Cast LayerNorm weight/bias to fp32 in the eager path (fixes
  half-precision eager fallbacks)
- Remove redundant logits write-back in _apply_logits_processors (torch
  rejects the self-assignment with a TP-padded vocab)
- Read Whisper STFT/mel/dither parameters from preprocessor_config.json
  instead of hardcoding; validate the audio window against
  max_source_positions at init

+++ tensorrt_llm/_torch/modules/layer_norm.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -89,11 +89,13 @@ class LayerNorm(nn.Module):
             hidden_states = hidden_states + residual.to(torch.float32)
             residual = hidden_states.to(input_dtype)
 
+        # Eager torch.layer_norm needs weight/bias in the fp32 compute dtype;
+        # torch.compile does this promotion implicitly (bitwise-identical).
         hidden_states = nn.functional.layer_norm(
             hidden_states,
             (hidden_states.shape[-1], ),
-            weight=self.weight,
-            bias=self.bias,
+            weight=self.weight.to(torch.float32),
+            bias=self.bias.to(torch.float32),
             eps=self.variance_epsilon,
         ).to(input_dtype)
